Add tests for null/empty jwt algorithms
If the appropriate supported algorithm option for DPoP, private_key_jwt,
or JAR are set to null or empty, we omit that from discovery. This is
non-standard behavior that we advise against in the xmldoc, but it is
consistent with historical behavior of IdentityServer.
